Update enums to support 64-bit types and improve backend handling
Adjust enums to differentiate between 32-bit and 64-bit types by updating type aliases and constants across the system. Refine backend type management by replacing `WGPUInstanceBackend_All` with `WGPUBackendType_Undefined`. Ensure compatibility and correctness for larger enumerations and backend configurations.

+++ demo/common/src/commonMain/kotlin/webgpu/HelloTriangle.kt
@@ -36,7 +36,7 @@ class HelloTriangleScene(val device: WGPUDevice, val renderingContextFormat: UIn
                 targetCount = 1u
                 targets = WGPUColorTargetState.allocateArray(scope, 1u) { index, structure ->
                     structure.format = renderingContextFormat
-                    structure.writeMask = WGPUColorWriteMask_All.toULong()
+                    structure.writeMask = WGPUColorWriteMask_All
                 }.let { ArrayHolder(it.handler) }
             }
